MultiDimDictionaries

Documentation for MultiDimDictionaries.

MultiDimDictionaries._shift_key

MultiDimDictionaries._shift_key — Function

_shift_key(key::Tuple, dims::Tuple, dim::Int, dim_key=1)

Either shift the key in dimension dim by dims[d], or if dim > length(key) then extend the key to length dim by padding with 1s and insert dim_key at dimension dim.

For example:

f = x -> x + 3

shiftkey(Tuple(2, 2, 2), [3, 3, 3], 2) -> Tuple(2, 5, 2)

shiftkey(Tuple(2, 2, 2), [3, 3, 3], 3) -> Tuple(2, 2, 5)

shiftkey(Tuple(2, 2, 2), [3, 3, 3], 4) -> Tuple(2, 2, 2, 1)

shiftkey(Tuple(2, 2, 2), [3, 3, 3], 4, "X") -> Tuple(2, 2, 2, "X")

shiftkey(Tuple(2, 2, 2), [3, 3, 3], 0) -> Tuple(1, 2, 5, 2)

shiftkey(Tuple(2, 2, 2), [3, 3, 3], 0, "X") -> Tuple("X", 2, 5, 2)

This is an "injection": https://en.wikipedia.org/wiki/Disjoint_union